Christopher Harris pointed out a mistake in our pathetic and cryptic field packager naming convention.
IFA_BINARY converts its payload to hex, but IFA_LLBINARY and IFA_LLLBINARY
does not convert it, the 'A' in these LL[L]BINARY field packagers indicates
that the length is encoded in ASCII ...

Due to the fact that we have many code out there that relies on IFA_LLBINARY
and IFA_LLLBINARY, we've added a new couple of files, contributed by
Mr. Harris: IFA_LLABINARY and IFA_LLLABINARY

These files are based on the 1.4.6 version of ISOFieldPackagers, we should
modify them using the new Encoders/Decoders.