OneShotChannelAdaptor connects and disconnects a channel for every message exchange.
Based on ChannelAdaptor I replaced the Sender and Receiver threads with a
configurable number of Worker threads and made each Worker have a Channel. The
number of workers bounds the number of concurrent connections to the remote
host. I rearranged some of the code and removed the isConnected method. I
believe that at least 98% of the code is from the original ChannelAdaptor.