Applied contributed changes and additions by Alexander Yakovlev - THANK YOU!
REQUIREMENTS SUMMARY:
--------------------------------------------

The simulator had to implement issuer side of the ISO connection
according to a specification based on ISO8583-1993 with a few
requirements which at the time I did not find to be supported by jPOS
"out of box": (R1) Bitmaps represented by characters of hex digits in
EBCDIC; (R1.2) Fields with field level bitmaps: 16 bits and 32 bits
bitmaps;

(R2) Variable length fields with bitmaps with a format id 1 byte
indicator intended to "version" format of the fields in the future
development of the product. As per the current specs, all such fields
have only one format id so there was no a need for functionality to
implement multiple formats and just a "place holder" for the data was
good enough solution. The indicator is the first physically sub-element
in the fields preceding bitmap. It is not "addressed" by the bitmap.

(R3) Fields with binary data represented in characters of hex digits in
EBCDIC example DE-52 (Pin Block);

CHANGES:
-------

(C1) Implemented IFE_BITMAP to satisfy (R1).
As I see IFE_BITMAP class was added recently by you, but I also needed
to satisfy also (R1.2) to support 16 and 32 bit bitmaps.
Also as a part of this change:
- Added IFE_BITMAPTest class, added the call in PackagerTestCase;
- Added  method public static byte[] bitSet2byte (BitSet b, int bytes)
in ISOUtil class  to return byte[] array with length equal to
specified in bytes parameter;
- Modified public static BitSet hex2BitSet (byte[] b, int offset, int
maxBits) of ISOUtil class to process no more data than necessary for
maxBits and to return bitmaps non longer than maxBits.

(C2) Implemented the "placeholder" for format id to satisfy (R2). To
keep it simple I implemented the format id as a header. Conceptually
the line of thought was if an ISO message can have a header, why
internal composite elements (the fields with subfields) cannot have
headers too??
- Added headerLength property in ISOBasePackager to keep the length of
a header in bytes;
- Modified pack and unpack methods of ISOBasePackager to package and
unpackage the headerLength bytes into header property of ISOMsg.
- Modified GenericPackager to support headerLength property in XML
packager configurations;
Example:
<isofieldpackager
  id="55"
  name="IC card system related data"
  length="256"
  class="org.jpos.iso.IFE_LLLBINARY"
  packager="org.jpos.iso.packager.GenericPackager"
  maxValidField="27"
  emitBitmap="true"
  bitmapField="0"
  firstField="0"
  headerLength="1">
- Added testGeneric93ebcdic method into PackagerTestCase class;

(C3) Implemented IDE_BINARY, added EbcdicHexInterpreter.