Reverted c94ff02f2 not working in Windows
In Windows, it seems we can't rename an open file, so the strategy used
to keep the file open at rename time doesn't work. Until we find a
work-around (may be detecting the OS to enable/disable this feature,
or using an emergency log to /tmp), I'm reverting the patch.