Added IFB_LLHEX (can be used on encrypted track2s)
We found situations were field 35 was repurposed to send an encrypted
track2, sometimes with proprietary encription. IFB_LLNUM is not suitable
in those situations because it just deal with BCD nibbles, not HEX.

IFB_LLHEX solves the problem.

The HexNibblesPrefixer is just a helper to deal with the variable length
encoding.