jPOS-106 use serializable internal object
The ChannelAdaptor use many internal flags (ready, reconnect, locks)
that are placed in the Space. We used simple 'Objects', but this
prevents us from using a persistent space (like JESpace).

Although the use of a persistent space with the ChannelAdaptor is
questionable (it's better to use a SAF), we changed these objects
so that they are serializable (Boolean and Date) now.