QConfig register its configuration with the NR
Q2 pushes Configuration objects to Configurable QBeans, but
in situations where non Q2 code has to be used (i.e. servlets)
this QConfig service allow us to 'pull' the configuration by
locating the Configuration object in the NameRegistrar.

We suggest to avoid using this when possible as the push configuration
is usually better.

[ci skip]