backward compatibility after PR67
PR67 (https://github.com/jpos/jPOS/pull/67) adds a nice feature,
if we call Logger.log with a LogEvent not associated with a LogSource,
then the Logger.log implementation picks the default system Logger (Q2).

This is good, but we have situations where we define a service and do
not assign a logger, and we expected not to issue any output.

In theory, the service should check if a logger is present, but
in practice, there are many out there, in production, so we added
a backward compatibility hack, using the new 'honorSourceLogger' flag.

If we create a LogEvent passing a LogSource parameter, even if the
LogSource is null, then we set the honorSourceLogger flag and the
behavior is the same as prior to PR67.

If, on the other hand, we call the LogEvent() or LogEvent(String)
constructor, then PR67 feature is enabled.