More aggressive thread creation
When the number of sessions is less than max-sessions, we used
to slowly create new threads in order to shape the load.

The initial assumption was that a system would need more sessions
during some periods of the day where load is greater, so if we had
configured 32 sessions, and a max of 128, every 32 transactions we
would check to see if an extra session would help.

In reality, what happens most of the time is that load spikes due to
a system slow-down (i.e. a remote issuer not responding, a database
being temporarily slow), and when it spikes, it really spikes, all
at once.

This little one-line change will be more aggressive creating these
additional threads, we check if an extra session is needed at the
beginning of every transaction.