Update ISOMsg.java
If a field doesn't exist in the message dont throw an ISOException, instead return null. An example is a field of type say XXX.YY. If XXX is not present don't throw an ISOException, instead return null. If you use a getString(String fpath) you dont get an ISOException but if you used getValue you would. Now both behave the same though the method signature is not changed.