Logger.log catches LogListener's Throwables
Typical LogListeners implementations don't throw exceptions, but when
they do, we risk propagating them to the caller, a situation very
difficult to debug.

With this change, if a LogListener throws an exception, we add it to the
LogEvent so that it gets picked by the next listener.

It's a good idea to use a reliable LogListener as the last one in the
chain.