New Card file will replace CardHolder
The class `org.jpos.core.CardHolder` has a wrong name, it's supposed
to represent a Card, but the name is confusing.

At some point, we will deprecate it and will replace it by this new
Card class, that in turn has support for Track1, Track2, pluggable LUHN
calculator. We plan to add support to generate track1 and track2 data.

This is s PoC class that may flux a bit for a while. Use at your own
risk for now.

The easiest way to learn how to use it is by looking at the `CardTest`
class.