Fixed ISOUtil.byte2BitSet(byte[] b, int offset, int maxBits) when 3 bitmaps (24 bytes) are requested, and the first bit is 0 and the 65th bit is 1.
If the first bit of the first byte is 0, only 1 bitmap (8 bytes) must be assumed.
Instead, if 192 bits were being requested, it looked at the 65th bit (as if a secondary bitmap were present), forgetting the fact
that the first bit was 0.

Also Fixed IFB_BITMAP.unpack (ISOComponent c, byte[] b, int offset) to consider the first bit, so everything works in
conjunction.

Commented out one test case in ISOUtilTest which was relying on the wrong behavior.