Builtin commands, colon to run subsystems, semicolon now escaped
We've recently introduced CLI 'subsystems', for example, one can type
'deploy' and enter the 'deploy' subsystem, with its commands.

Subsystems have to have a way to exit and come back to the top-level
shell, so they need an EXIT implementation. Because EXIT is required in
all subsystems, we've added a special package 'org.jpos.q2.cli.builtin'
to hold builtin commands, available in all subsystem.

This commit also adds a handy feature to run one-shot subsystem commands
and get back to the main shell, for example, instead of typing:

  q2> deploy
  deploy> list
  deploy> exit
  q2>

one can run 'deploy:list'

This commit also adds support for multi-line commands (line must end with the '\' character)
and add support to escape a semi-colon using a backlash in order to tell
apart semicolons that intend to separate commands versus those that
actually want to use a semicolon as a parameter, i.e.:

  q2> echo a\;b; echo c

will output:

   a;b
   c