defaults for commit, abort and prepareForAbort
We see a lot of transaction implementations having to override commit
and rollback just to honor the API. Same happens for participants that
want to be engaged at abort time so they need to implement
AbortParticipant, but the prepareForAbort implementation just calls
prepare.

We now provide default implementations for those.