handle debug logging on TMs without participants
While it makes no sense to have a TM without participants, during tests
and initial setup, it is possible to have one, and we still want it to
log the Context it's receiving.

This commit also cleans up a bit the log and persistent space location
used for testing (we were using a 'target' directory from old Maven
builds).