handy 'FAIL' combination
Returning ABORTED | READONLY | NO_JOIN is a popular combination whenever
a failure occurs in a participant. ABORTED happens to be equal to 0 so
the smart IDE always warns that it could be ommitted, but if we
omit it, the code looses a good indication that we are aborting.

Added this 'FAIL' constant as a trade-off.