Better handling of emergency thread creation
When the configuraiton parameter 'max-sessions' is greater than
'sessions', the TransactionManager is supposed to go up and down
in the number of threads (sessions) it uses.

Before this change, the main sessions were in charge of creating
additional sessions, but if all main sessions were locked for
some reason waiting for its participants to finish, no extra
sessions were created.

We now use a ScheduledThreadPoolExecutor to regularly check
sessions/max-sessions/outstanding and decide if new sessions
need to be created.