issue #169 - Manifest classpath in artifact
jPOS is both a library, and a framework. Traditionally, we used
to run the jPOS framework off the jPOS jar (using `java -jar jpos.jar`).

While this is a valid use when using Q2 and our own directory conventions
(i.e. dependencies in the 'lib' directory), this is not right for people
using jPOS as a library, pulling the dependency from Maven Central as properly
reported by @narwajea.

With this change, the jpos main jar (the one published to Maven) won't contain
a `Class-Path` nor a `Main-Class`. In order to keep our handy ability (mostly for
development/testing purposes) to run Q2 from the source tree using 'gradle installApp',
we now have a `q2jar` target, and in that case, we are producing a jpos-x.x.x-q2.jar
instead.