Update ISOUtil.java
Provide a protect method with a char to protect with. For backwards compatibility the underscore char ('_')is used. Useful when you want to output a pan which is compliant in terms of its packager. e.g. a prod settlement file pans needs to be protected and transferred to QA and still be usable.