If card is optional, but present, it has to be valid
putCard was raising a 'fail' if card was not present even though
it was optional. Now, if the card is optional and not present, then
no failure is raised. On the other hand, if card is optional, and
present, we check it's also valid, otherwise we raise a fail.

In other words, card can be optional, but if present, it has to be a good one.