Configurable prepend machine name to log file name
Added the ability to prepend the machine name to the log file name to easily identify logs when brought to a central location from multiple servers.
This change is backwards compatible and will prepend the machine based on the configuration in the log listener. By default, prepending machine name is set to false.
An optional configurable separator after the prepended machine name gives some flexibility to naming conventions.