File name token replacement now depends on only the existence of a "file-name-pattern" property.
The file-name-pattern property defines the tokens to replace in the "file" property.

Currently the only 2 tokens supported are the following:
1) h - Replace %s with current hostname.
2) e{XXX} - replace %s with environment variable with key XXX.

Added runPostConfiguration default method to the Configurable interfaces.
This is supposed to be used to split configuration from logic that needs
to execute after configuration, making inheritence of configuration and
logic less complicated.

The QFactory has been modified to call this method for both Configurable and XmlConfigurable.

DailyLogListener has been modified to call the setConfiguration method of its parent at a time it
makes sense.  This is tricky since in some cases the parent must cater for possible overrides.

The logic for scheduling a timer and rotation of files has been abstracted away in closures
which are configured at setConfiguration time.

This allows DailyLogListener to inherit both the runPostConfiguration and logRotate methods
of its parent, thus DailyLogListener gains rotate-on-start and file-name-pattern by default.