Optimize toArray() calls
Calls to a collection’s toArray(E[]) method should specify a target array of zero size.
This allows the JVM to optimize the memory allocation and copying as much as possible.

For more information see:
https://pmd.github.io/latest/pmd_rules_java_performance.html#optimizabletoarraycall

In this case is even better to call toArray() (which is done inside SimpleMsg)
when collection is passed.