FlatLogListener that uses RemoveNewLinesMapper to remove newlines and optional combine spaces in LogEvents.
This differs from using RemoveNewLinesMapper in the MappingLogEventWriter in that it modifies the LogEvent returning a FrozenLogEvent instead of modifying the data just before writing to the output.

This listener thus doesn't write to output, but similarly to the ProtectionLogListener, modifies the LogEvent that is processed by subsequent LogListeners defined in the LoggerAdapter.