Ensure StringBuilder exceptions are deterministic on Java 15.
It appears the NullPointerException exception message is
non-deterministic when passing null to StringBuilder under
Java 15 due to what appears to be a race of some sort likely
due to internal implementation changes in StringBuilder.

Fix this by checking for and raising our own NullPointerException
before null gets passed to StringBuilder.