feat: Add Interfaces for application cryptogram spec
The different implementations (versions) of the application cryptogram
generation algorithm differ in 4 points: the master key derivation
method, the session key derivation method, the data elements used and
the generation method of the response cryptogram.

To create components that can generate and validate the application
cryptograms in a generic way, we need a mechanism to specify the
particularities of the different algorithm.

Interfaces are added to generically represent the different cryptogram
generation algorithms of the application. Additionally, an interface is
created that represents the set of data used by the different
algorithms to generate the cryptogram.