feat: Add spec for VISA cryptogram generation
Visa uses several versions of the application cryptogram
(CVN 10, 18 and '22' ), each version has its own peculiarities.

We need to specify the different versions of the application cryptogram
generation algorithm used by VISA so that the generic components (SM)
in charge of generating and validating the cryptogram can correctly
perform their task.

The specification of the cryptogram generation algorithm of the
application used by VISA is added, this specification covers all
versions and particularities.