feat: Add support to detect criptogram spec
The IAD contains application data, among these data is
CVN (Cryptogram version number), this data indicates the version of
the algorithm used to generate the cryptogram.

In environments where several versions of the application cryptogram
coexist, we need to automatically detect the specifications of the
algorithm used to generate the current cryptogram.

Support is added to detect the specifications of the cryptogram
generation algorithm, using the format and CVN information contained
in IAD.