PAUSE is back, for backward compatibility
With project Loom, TM continuations are not required.

In order to support participants that used to return PAUSE and have not
been migrated to synchronous mode, the Context now implements a new
Pausable interfase (with pause/resume methods).

When a participant returns PAUSE, the TM actually waits until the
context gets resumed.