ISOServer/ChannelAdaptor default to Platform Threads.
During 'jinstress' test we found that for persistent connections, the
ChannelAdaptor and ISOServer performs better using platform threads.
There are situations, such as a massive number of low-traffic POS
devices that could be better served using VirtualThreads, so there's
still the ability to set 'virtual-threads' to 'true' in the components'
configuration.