Temporarily remove SSHd support due to multiple CVEs
The embedded SSHd support has been temporarily removed due to security
concerns, primarily stemming from its dependency on the
net.i2p.crypto:eddsa library, which is affected by the long-standing
CVE-2020-36843 vulnerability. Although this does not pose a direct
risk—since the SSH support is only launched when Q2 is started with the
rarely used --ssh option—the presence of this vulnerable dependency
triggers red flags in every PCI audit.

Originally introduced as an experimental feature in jPOS-EE, SSHd
support was later migrated to the jPOS core, a decision that in
hindsight added unnecessary complexity and risk to the base
distribution. Given its limited usage (mainly for demonstration
purposes), and its current cost in terms of compliance overhead,
removing it from the core codebase is the most prudent step.

This change aims to streamline PCI DSS audits by eliminating the
dependency to org.apache.sshd:sshd-core and its dependency
net.i2p.crypto:eddsa.

We will revisit this feature in the future to evaluate whether it
warrants reintroduction—potentially in a more modular and secure fashion
(i.e. back to jPOS-EE)—or if alternative mechanisms can meet the same
operational needs.